requirements.txt
Flask==3.1.2
opencv-python==4.12.0.88
imutils==0.5.4
numpy==2.2.6
pillow==11.3.0
## The following requirements were added by pip freeze:
blinker==1.9.0
click==8.2.1
colorama==0.4.6
colormath==3.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
networkx==3.5
scikit-learn==1.7.2
scipy==1.16.1
threadpoolctl==3.6.0
Werkzeug==3.1.3